refactor [entity,mapper] id数据类型统一为Long/bigint

+++ meeting-appointment-modules/src/main/java/com/jfzt/meeting/entity/vo/MinutesPlanVO.java
@@ -19,7 +19,7 @@ public class MinutesPlanVO {
     /**
      * 会议纪要id
      */
-    private Integer minutesId;
+    private Long minutesId;
     /**
      * 迭代内容
      */
